Update FMain.pas
rincipal change is the procedure to generate automatic id nos for resources. Also includes changes to openFile, ClearDocument, ActionSaveAsExecute and ActionSaveExecute. If a resource name other than "APPLICATION" is used, it needs to be saved. I tried to save it with the ribbon xml, but of course it offended the schema. So it is saved with in an ini file with the same name as the xml.

BuildAndPreview has been changed to pass in an additional parameter (ie ResourceName) for the preview form and the compiler.